requirements.txt
joblib==1.0.1
numpy==1.21.2
Pillow==8.3.2
scikit-learn==0.24.2
scipy==1.7.1
threadpoolctl==2.2.0
